fix: Correct hook call order in MemberModal
Moves the `useMemo` hook for `currentPartnerNames` in
`src/components/MemberModal.tsx` to the top level of the functional
component.

This change ensures that all hooks are called in the same order on every
render, preventing the "Rendered more hooks than during the previous
render" (React Error #310). The `parsePartnerString` function used
within this `useMemo` is designed to handle potentially null or
undefined inputs that might occur before `detailedMember` is fully loaded,
making this move safe.

This fix should resolve the React Error #310 and may also help
stabilize the component, potentially resolving other downstream errors.
